Demos/Azure Durable Functions Demo - Fan-Out Fan-In/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

#TODO: Make sure this package is always added when using Python durable functions
azure-functions-durable

numpy 